\name{S_downward}
\alias{S_downward}
\title{S_downward when status is passed}
\usage{
S_downward(n)
}
\arguments{
  \item{n}{A vector containg number of people in each observation}
}

\description{
Takes input parameter as n and returns the value of downward looking status, i.e. sd.
}
\examples{
S_downward(c(0,50,25,25))
}
