\name{S_upward}
\alias{S_upward}
\title{S_upward when status is passed}
\usage{
S_upward(n)
}
\arguments{
  \item{n}{A vector containg number of people in each observation}
}

\description{
Takes input parameter as n and returns the value of upward looking status, i.e. sd.
}
\examples{
S_upward(c(0,50,25,25))
}
